add example for summing over list

+++ src/main/kotlin/com/github/navyazaveri/dynamik/stdlib/List.kt
@@ -25,8 +25,8 @@ class ListInstance : ContainerInstance() {
         return _list.toString()
     }
 
-    override fun len(): Int {
-        return _list.size
+    override fun len(): Double {
+        return _list.size.toDouble()
     }
